Add files via upload
The backend code here is written in python. All the dependencies written as imports in the App_integrated_backend.py file needs to be added to the python environment. Once those dependencies are installed, run the App_integrated_backend.py file. Then launch the application and the connection should be established between the server and the app.